hopefully Cam works now

+++ code/code/Gps.hpp
@@ -18,8 +18,8 @@
 struct GnssData {
 
   // positional variables
-  float lat = 0.0;
-  float lon = 0.0;
+  double lat = 0.0;
+  double lon = 0.0;
   float alt = 0.0;
 
   // num of satellites connected
